pyproject.toml
[tool.poetry]
name = "pe-fasttext"
version = "0.1.0"
description = "Positionally-Enhanced FastText embeddings for biological sequences"
authors = ["Michael Bohl <mibohl@student.ethz.ch>", "Tudor-Stefan Cotet <tcotet@student.ethz.ch>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "pe_fasttext", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
gensim = "^4.3.0"
numpy = "^1.26.0"
scikit-learn = "^1.3.0"
faiss-cpu = "^1.7.4"
PyYAML = "^6.0.0"
tqdm = "^4.66.0"
modal = "^0.57.2"
requests = "^2.31.0"
biopython = "^1.83"
torch = {version="^2.1.0", optional=true}
datasets = "^2.18.0"
mg_toolkit = "*"

[tool.poetry.extras]
# Torch is optional because it is only needed for RoPE (complex mul) or GPU acceleration.
gpu = ["torch"]

[tool.poetry.scripts]
peft-download = "pe_fasttext.cli:download_datasets_cli"
peft-train = "pe_fasttext.cli:train_fasttext_cli"
peft-embed = "pe_fasttext.cli:embed_cli"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api" 